+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- ModelService.GetComputeTaskInputModels, use ComputeTaskAPI.GetInputAssets instead
+
 ## [0.25.0] - 2022-08-29
 
 ### Added